requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
redis==4.5.5
pillow==9.5.0
python-multipart==0.0.6
pydantic==1.10.7
aiohttp==3.8.4
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
six==1.16.0 